Added mongodb
Created a mongodb database in mlab. Added the references in config/prod.json and executed with NODE_ENV=prod. The mongodb has two collections frontend and backend, which are prepopulated with values for two attributes version and env. So, the app just read those values and then prints them in the response.